manga dtos update

+++ src/main/java/br/com/clone/unionmangas/enumerator/AgeGroupEnum.java
@@ -2,24 +2,24 @@ package br.com.clone.unionmangas.enumerator;
 
 public enum AgeGroupEnum {
 
-    LIVRE(0L, "LIVRE", "Livre para todos os publicos"),
-    MAIOR_10(10L, "MAIOR_10", "Não recomendado para menores de 10 anos"),
-    MAIOR_12(12L, "MAIOR_12", "Não recomendado para menores de 12 anos"),
-    MAIOR_14(12L, "MAIOR_14", "Não recomendado para menores de 14 anos"),
-    MAIOR_16(12L, "MAIOR_16", "Não recomendado para menores de 16 anos"),
-    MAIOR_18(12L, "MAIOR_18", "Não recomendado para menores de 18 anos");
+    LIVRE(0, "LIVRE", "Livre para todos os publicos"),
+    MAIOR_10(10, "MAIOR_10", "Não recomendado para menores de 10 anos"),
+    MAIOR_12(12, "MAIOR_12", "Não recomendado para menores de 12 anos"),
+    MAIOR_14(12, "MAIOR_14", "Não recomendado para menores de 14 anos"),
+    MAIOR_16(12, "MAIOR_16", "Não recomendado para menores de 16 anos"),
+    MAIOR_18(12, "MAIOR_18", "Não recomendado para menores de 18 anos");
 
-    private Long age;
+    private Integer age;
     private String chave;
     private String description;
 
-    AgeGroupEnum(Long age, String chave, String description) {
+    AgeGroupEnum(Integer age, String chave, String description) {
         this.age = age;
         this.chave = chave;
         this.description = description;
     }
 
-    public Long getAge() {
+    public Integer getAge() {
         return age;
     }
